+++ CheckYourEligibility-Admin/Controllers/BaseController.cs
@@ -1,19 +1,14 @@
-using CheckYourEligibility_DfeSignIn;
 using CheckYourEligibility_DfeSignIn.Models;
-using CheckYourEligibility_FrontEnd.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using System.Diagnostics;
-using System.Security.Claims;
 
 namespace CheckYourEligibility_FrontEnd.Controllers
 {
     [Authorize]
     public class BaseController : Controller
     {
-       
-       protected DfeClaims? _Claims;
+
+        protected DfeClaims? _Claims;
 
     }
 }
