+++ CheckYourEligibility-Admin/Views/BulkUpload/Batch_Check.cshtml
@@ -11,7 +11,7 @@
     <partial name="_ValidationSummary" model="ViewData.ModelState" />
 }
 
-<form asp-controller="School" asp-action="Batch_Check" method="post" enctype="multipart/form-data">
+<form asp-controller="BulkUpload" asp-action="Batch_Check" method="post" enctype="multipart/form-data">
     <div class="govuk-grid-row">
         <div class="govuk-grid-column-two-thirds">
             <h1 class="govuk-heading-l">Check free school meals – multiple students </h1>
@@ -30,7 +30,7 @@
             </div>
 
             <button class="govuk-button" data-module="govuk-button">
-                Upload
+                Perform Checks
             </button>
         </div>
     </div>
