--- CheckYourEligibility-Admin/Views/School/BatchOutcome/Default.cshtml
@@ -1,1 +1,0 @@
-﻿Test
