--- CheckYourEligibility-Admin/Views/School/BatchOutcome/Success.cshtml
@@ -1,17 +1,0 @@
-﻿<div class="govuk-notification-banner govuk-notification-banner--success" role="alert" aria-labelledby="govuk-notification-banner-title" data-module="govuk-notification-banner">
-    <div class="govuk-notification-banner__header">
-        <h2 class="govuk-notification-banner__title" id="govuk-notification-banner-title">
-            Success
-        </h2>
-    </div>
-    <div class="govuk-notification-banner__content">
-        <h1 class="govuk-notification-banner__heading">
-            Checks completed
-        </h1>
-        <p class="govuk-body">Now download the checked file to see which parents can claim free school meals for their children.</p>
-        <p>free-school-meal-outcomes-YYYMMDD.csv</p>
-
-         @Html.ActionLink("Download", "Batch_check_download", "School", new { @class = "govuk-button" })
-
-    </div>
-</div>
